feat: Add comprehensive test data population script and README for test users. Introduce test user credentials and partnerships, along with habit and goal creation for testing purposes. Enhance user search functionality in the frontend with debounced input handling.

+++ Backend/app/models/goals.py
@@ -58,7 +58,7 @@ class UserGoal(BaseModel):
     )
 
     # Progress tracking (moved from habit level to per-user)
-    goal_progress: int = Field(default=0, ge=0, description="Current progress (count for frequency, accumulated value for completion)")
+    goal_progress: float = Field(default=0, ge=0, description="Current progress (count for frequency, accumulated value for completion)")
     count_checkins: int = Field(default=0, ge=0, description="Number of check-ins completed")
     checked_in: bool = Field(default=False, description="Checked in for current period")
 
@@ -184,7 +184,7 @@ class UserGoalResponse(BaseModel):
     target_value: Optional[float]
 
     # Progress
-    goal_progress: int
+    goal_progress: float
     count_checkins: int
     total_checkins_required: Optional[int]
     progress_percentage: float
